Use MessageManager instead of QMessageBox for showing messages
QMessageBox shows a modal message box, which is quite intrusive. For
instance if you run 2 instances of Qt Creator, then the second one will
start with a message box saying that it couldn't listed to the TCP port.
Instead if we use the MessageManager class to send error messages the
General Messages pane will flash to notify the user that a message is
there, but otherwise not bother the user.